Create test of notification model
